submission/requirements.txt
python-dotenv
langgraph
langgraph-sdk
langgraph-checkpoint-sqlite
langsmith
langchain-community
langchain-core
langchain-openai
langgraph-supervisor
langgraph-swarm 
langchain-tavily
langchain-anthropic
langchain-sandbox
langchain_experimental
notebook
tavily-python
trustcall
langgraph-cli[inmem]
matplotlib
numpy
seaborn
reportlab
scipy
networkx
wordcloud
scikit-learn
statsmodels
textblob
geopandas
fpdf
plotly
pandas
lifelines
mlxtend
prophet
xgboost